fix: zip64 support for large dictionaries and import dialog timing
- Add Zip64 extra field parsing for large archives (>4GB offsets/sizes)
- Add bounds check before reading compressed data to prevent out-of-bounds
- Fix dictionary import dialog to show after file picker returns, not during

+++ hibiki/android/app/src/main/cpp/hoshidicts_src/zip/zip.hpp
@@ -10,9 +10,9 @@
 struct ZipEntry {
   std::string name;
   uint16_t compression_method;
-  uint32_t compressed_size;
-  uint32_t uncompressed_size;
-  size_t data_offset;
+  uint64_t compressed_size;
+  uint64_t uncompressed_size;
+  uint64_t data_offset;
 };
 
 struct Zip {
